refactor: wire add-on execution through __init__.py
